added fast and pacman states for snake
