+++ src/main/assets/snake/normalState.py
@@ -1,0 +1,5 @@
+from assets.snake.snakeState import SnakeState
+
+class NormalState(SnakeState):
+    def update(self):
+        super().update()
