+++ src/main/assets/snake/snakeState.py
@@ -1,0 +1,15 @@
+class SnakeState:
+    def __init__(self, snake):
+        self.snake = snake
+
+    def handle_input(self, event):
+        pass
+
+    def update(self):
+        pass
+
+    def on_enter(self):
+        pass
+
+    def on_exit(self):
+        pass
