+++ src/main/state/PlayingState.py
@@ -3,7 +3,7 @@ import time
 from pygame.math import Vector2
 from state.GameState import GameState
 from assets.redapple import RedApple
-from assets.snake import Snake
+from assets.snake.snake import Snake
 from phases.LevelManager import LevelManager
 import os
 from config import LEVEL_DIR, CELL_SIZE, SCREEN_HEIGHT, SCREEN_WIDTH
