+++ src/main/state/PlayingState3.py
@@ -4,7 +4,7 @@ from pygame.math import Vector2
 from state.GameState import GameState
 from assets.redapple import RedApple
 from assets.Hole import Hole
-from assets.snake import Snake
+from assets.snake.snake import Snake
 from phases.LevelManager import LevelManager
 import os
 from config import LEVEL_DIR, CELL_SIZE, SCREEN_HEIGHT, SCREEN_WIDTH
